Revert "Removing useless comments"
This reverts commit 3097946916ba158bbc736896a412519453fe4614.

The code is made to be read by a human first. Then by a machine.
There is very little documentation in the entire project, which makes it difficult to maintain for outsiders.

Example: The sign of a literal that is false if the value is positive is very clearly counter-intuitive and is not documented in the Lit class.
Documentation including examples of use is crucial.
The self-documented code is a legend and does not exist.